reverted index in favor of index.web
