+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.10",
   "description": "Stack navigator component for React Navigation",
   "main": "dist/index.js",
-  "module": "src/index.js",
+  "module": "src/index.web.js",
   "sideEffects": false,
   "files": [
     "dist/",
